Moved data_types into const and other small changes
