+++ amulet_nbt/amulet_nbt_py/const.py
@@ -1,0 +1,4 @@
+SNBTType = str
+
+CommaNewline = ",\n"
+CommaSpace = ", "
