--- amulet_nbt/amulet_nbt_py/data_types.py
@@ -1,1 +1,0 @@
-SNBTType = str
